tests/submission/Ms2ger/obsolete-features/requirements-for-implementations/other-elements-attributes-and-apis/document.all-02.html

<!DOCTYPE html>
<title>document.all: length</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="http://www.whatwg.org/html5/#dom-document-all">
<link rel="help" href="http://www.whatwg.org/html5/#htmlallcollection-0">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  assert_true(!!document.all.length);
});
</script>
